requirements.txt
python-dotenv
pinecone
openai